Switch tiers from item-count to cumulative-revenue based
Threshold ladders (both a group's own tiers and a standalone item's) now
compare against cumulative BASE (pre-multiplier) money earned selling
that category/item, not a raw item count. E.g. "threshold: 10000" now
means "$10,000 earned", not "10,000 items sold". Config key/threshold
numbers are unchanged - only their meaning is.

Design decision: revenue counted is the sale's base price, not the
boosted amount actually paid - avoids a compounding feedback loop and
keeps pacing predictable regardless of a player's current multiplier.
Base revenue accrues at an exactly constant rate per unit, so the ledger
update itself needs no iterative stepping - only the price actually
charged does (since the multiplier can still change mid-sale as revenue
crosses a threshold partway through the stack sold).

PlayerDataManager gains parallel group/item revenue tracking (data.yml),
independent of the existing item-count "sold" map, which keeps working
unchanged for /pm stats. ConfigManager gains a revenue-based method
family (revenueMultiplierAt/revenueActiveThreshold/
revenueNextThresholdAbove/computeTieredRevenueSaleValue/
computeUnifiedRevenueSaleValue) alongside the untouched count-based ones.
SellProcessor, MilestoneManager, MenuManager, ProfitPlaceholders, and
/pm stats all updated to read/display revenue instead of count.

ProfitMultiplierAPI gains getGroupRevenue/getGroupMultiplier/
getItemRevenue/getItemMultiplier; the old count-based multiplier/
threshold/calculateSaleValue methods are marked @Deprecated (no longer
reflect real pricing) rather than removed, since this is a published API
jar. ThresholdReachedEvent's newTotal/threshold fields changed long ->
double (an intentional breaking signature change - they're a currency
amount now).

+++ src/main/java/me/docdrewskii/profitmultiplier/api/events/ThresholdReachedEvent.java
@@ -11,13 +11,18 @@ public class ThresholdReachedEvent extends Event {
 
     private final Player player;
     private final Material material;
-    private final long newTotal;
+    private final double newTotal;
     private final double previousMultiplier;
     private final double newMultiplier;
-    private final long threshold;
+    private final double threshold;
 
-    public ThresholdReachedEvent(Player player, Material material, long newTotal,
-                                 double previousMultiplier, double newMultiplier, long threshold) {
+    /**
+     * @param newTotal cumulative BASE revenue (money, not item count) earned on this ladder
+     *                 after the sale that crossed the tier
+     * @param threshold the (currency) threshold that was just crossed
+     */
+    public ThresholdReachedEvent(Player player, Material material, double newTotal,
+                                 double previousMultiplier, double newMultiplier, double threshold) {
         this.player = player;
         this.material = material;
         this.newTotal = newTotal;
@@ -34,7 +39,7 @@ public class ThresholdReachedEvent extends Event {
         return material;
     }
 
-    public long getNewTotal() {
+    public double getNewTotal() {
         return newTotal;
     }
 
@@ -46,7 +51,7 @@ public class ThresholdReachedEvent extends Event {
         return newMultiplier;
     }
 
-    public long getThreshold() {
+    public double getThreshold() {
         return threshold;
     }
